+++ app/src/main/res/values/ids.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <item name="button" type="id" />
+    <item name="mamagebtn" type="id" />
 </resources>
